requirements.txt
streamlit
groq
python-dotenv
fpdf2
streamlit-echarts